@@ -10,6 +10,7 @@ import Combine
 
 extension UIControl {
 
+    @MainActor
     class InteractionSubscription<S: Subscriber>: Subscription where S.Input == Void {
 
         private let subscriber: S?
@@ -28,12 +29,12 @@ extension UIControl {
             _ = self.subscriber?.receive(())
         }
 
-        func request(_ demand: Subscribers.Demand) {}
+        nonisolated func request(_ demand: Subscribers.Demand) {}
 
-        func cancel() {}
+        nonisolated func cancel() {}
     }
 
-    struct InteractionPublisher: Publisher {
+    struct InteractionPublisher: @preconcurrency Publisher {
 
         typealias Output = Void
         typealias Failure = Never
@@ -46,6 +47,7 @@ extension UIControl {
             self.event = event
         }
 
+        @MainActor
         func receive<S>(subscriber: S) where S : Subscriber, Never == S.Failure, Void == S.Input {
             let subscription = InteractionSubscription(
                 subscriber: subscriber,
