@@ -6,6 +6,7 @@
 //
 
 import SwiftUI
+import Factory
 
 // MARK: - View Controller Extension
 
@@ -13,13 +14,14 @@ extension UIViewController {
 
     open override func motionEnded(_ motion: UIEvent.EventSubtype, with event: UIEvent?) {
         if motion == .motionShake {
-            let viewController = AppDebugModeProvider.shared.start()
-
-            if presentedViewController == nil, view.window?.rootViewController?.presentedViewController == nil {
-                viewController.modalPresentationStyle = .fullScreen
-                present(viewController, animated: true)
-            } else {
-                dismiss(animated: true)
+            Task { 
+                let viewController = await Container.shared.packageManager.resolve().start()
+                if presentedViewController == nil, view.window?.rootViewController?.presentedViewController == nil {
+                    viewController.modalPresentationStyle = .fullScreen
+                    present(viewController, animated: true)
+                } else {
+                    dismiss(animated: true)
+                }
             }
         }
     }
