@@ -9,14 +9,15 @@ import SwiftUI
 
 enum AppDebugColors {
     
-    static let backgroundPrimary = Color(hex: "#000000")
-    static let backgroundSecondary = Color(hex: "#282833")
-    static let primary = Color(hex: "#CFE838")
-    static let secondary = Color(hex: "#1F0CF2")
-    
-    static let black = Color(hex: "#111117")
-    
-    static let textPrimary = Color(hex: "#FFFFFF")
-    static let textSecondary = Color(hex: "#A5ABBA")
-    
+    static let primary = Color(#colorLiteral(red: 0.8445237279, green: 0.9138714671, blue: 0.2786209583, alpha: 1))
+    static let secondary = Color(#colorLiteral(red: 0.1686843038, green: 0.2027454674, blue: 0.9607277513, alpha: 1))
+
+    static let textPrimary = Color(#colorLiteral(red: 1, green: 0.9999999404, blue: 1, alpha: 1))
+    static let textSecondary = Color(#colorLiteral(red: 0.7056729198, green: 0.7281373143, blue: 0.7778759599, alpha: 1))
+
+    static let backgroundPrimary = Color(#colorLiteral(red: 0, green: 0, blue: 0, alpha: 1))
+    static let backgroundSecondary = Color(#colorLiteral(red: 0.2080399692, green: 0.2099102139, blue: 0.2606674135, alpha: 1))
+
+    static let black = Color(#colorLiteral(red: 0.08346965164, green: 0.08476512879, blue: 0.1187344268, alpha: 1))
+
 }
