@@ -6,6 +6,6 @@ final class AppDebugModeTests: XCTestCase {
         // This is an example of a functional test case.
         // Use XCTAssert and related functions to verify your tests produce the correct
         // results.
-        XCTAssertEqual(AppDebugMode().text, "Hello, World!")
+//        XCTAssertEqual(AppDebugMode.text, "Hello, World!")
     }
 }
